refactor(wiki): move ChangeKind/CommitResult/AiRebaseMaxRetriesError to app/wiki/models.py
Follows the domain-level models.py pattern (app/ingest/models.py,
app/web/models.py). Removes these from app/models/file_system.py
(HTTP shapes only) and app/wiki/types.py (deleted).
