+++ backend/app/tasks/triggers.py
@@ -57,7 +57,7 @@ from app.triggers.engine import (
 )
 from app.wiki import acl as wiki_acl
 from app.wiki import git as wiki_git
-from app.wiki.types import ChangeKind
+from app.wiki.models import ChangeKind
 
 log = logging.getLogger(__name__)
